RoadQuality/requirements.txt
# Hardware interface
pyserial
pynmea2
fastestrplidar
smbus2

# Data processing and visualization
matplotlib
numpy
folium
opencv-python

# For GPX export (if you use this functionality)
gpxpy
pandas